api/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos
aiohttp
python-dotenv
fastapi==0.100.1
uvicorn[standard]==0.23.2
